backend/tests_requirements.txt
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
blowfish==0.6.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
exceptiongroup==1.0.4
fastapi==0.86.0
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
msgpack==1.0.4
neovim==0.3.1
packaging==22.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pynvim==0.4.3
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
